csds_array.h: dynamic array (with abstracted header) first steps
This new dynamic array implementation intends to replace the Vec struct
already present in the library. It uses a syntax similar to what you
find with 'stb arr´ and uses a single header file. Also, it will support
any type without changing a macro definition like previously, allowing the
use of multiple arrays of more than one type in a C file. For now Vec will
still be on the repository until array is implemented and stable enough.